refactor: reorganize import entries
